+++ lib/wine_bouncer/auth_strategies/protected.rb
@@ -1,34 +1,34 @@
 module WineBouncer
   module AuthStrategies
-    class Protected
+    class Protected < WineBouncer::BaseStrategy
 
-      def endpoint_protected?(context)
-        has_authorizations?(context)
+      def endpoint_protected?
+        has_authorizations?
       end
 
-      def has_auth_scopes?(context)
-        endpoint_authorizations(context) && 
-          endpoint_authorizations(context).has_key?(:scopes) && 
-          endpoint_authorizations(context)[:scopes].any?
+      def has_auth_scopes?
+        endpoint_authorizations &&
+          endpoint_authorizations.has_key?(:scopes) &&
+          endpoint_authorizations[:scopes].any?
       end
 
-      def auth_scopes(context)
-        endpoint_authorizations(context)[:scopes].map(&:to_sym)
+      def auth_scopes
+        endpoint_authorizations[:scopes].map(&:to_sym)
       end
 
       private
 
-      def nil_authorizations?(context)
-        endpoint_authorizations(context).nil?
+      def nil_authorizations?
+        endpoint_authorizations.nil?
       end
 
 
-      def has_authorizations?(context)
-        nil_authorizations?(context) || endpoint_authorizations(context)
+      def has_authorizations?
+        nil_authorizations? || endpoint_authorizations
       end
 
-      def endpoint_authorizations(context)
-         context[:auth]
+      def endpoint_authorizations
+        api_context.options[:route_options][:auth]
       end
     end
   end