+++ spec/dummy/app/api/protected_api.rb
@@ -34,6 +34,10 @@ module Api
     get '/oauth2_dsl' do
       { hello: 'oauth2_dsl' }
     end
+
+    get '/not_described_world' do
+      { hello: 'non described world' }
+    end
   end
 
   class ProtectedApiUnderTest < Grape::API