+++ spec/intergration/oauth2_swagger_strategy_spec.rb
@@ -98,6 +98,16 @@ describe Api::MountedSwaggerApiUnderTest, type: :api do
     end
   end
 
+  context 'not_described_world' do
+    it 'allows to call an endpoint without description' do
+      get '/swagger_api/not_described_world'
+      expect(last_response.status).to eq(200)
+      json = JSON.parse(last_response.body)
+      expect(json).to have_key('hello')
+      expect(json['hello']).to eq('non described world')
+    end
+  end
+
   context 'resource_owner' do
     it 'is available in the endpoint' do
       get '/swagger_api/protected_user', nil, 'HTTP_AUTHORIZATION' => "Bearer #{token.token}"